python_backend/requirements.txt
fastapi
uvicorn
opencv-python-headless
ultralytics
requests
python-dotenv
numpy
